Refactoring
* Moved away from large monlithic files.
* Split models out into separate files.
* Finished adding features to data layer.
* Added `rejectEmptyFields` util function.
* Cleaned up some code.
* Changed view code to take full advantage of the data layer.
